infra/assets/extra-requirements.txt
# Azure Functions core
azure-Functions

# GitHub Copilot SDK
github-copilot-sdk

# Pydantic for tool parameter validation
pydantic>=2.0.0

# azure-cli